requirement.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
django-tastypie==0.14.0
djangorestframework==3.7.1
idna==2.6
kombu==4.1.0
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
